pyproject.toml
[project]
name = "aci-agents"
version = "0.1.0"
description = "Agents and examples built with the ACI SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["aci-sdk>=0.0.2b2"]

[dependency-groups]
dev = [
    "anthropic>=0.49.0",
    "dotenv>=0.9.9",
    "crewai>=0.108.0",
    "langchain>=0.3.21",
    "langchain-openai>=0.3.11",
    "llama-index>=0.12.26",
    "openai>=1.68.2",
    "rich>=13.9.4",
    "ruff>=0.11.2",
    "mypy>=1.15.0",
]
